FarmLand-Backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
psycopg2-binary
sqlalchemy-utils
python-dotenv
flask-jwt-extended
flask-cors